Shortest remaining time strategy

+++ operation-systems/scheduler/src/main/kotlin/ru/chudakov/scheduling/SchedulingStrategy.kt
@@ -4,5 +4,5 @@ import ru.chudakov.Process
 import java.util.*
 
 interface SchedulingStrategy {
-    fun next(processes: LinkedList<Process>): Int
+    fun next(processes: LinkedList<Process>): Double
 }
